Feature/ssor-194-Dockerization

+++ boss-transaction-service/src/main/java/org/ssor/boss/transactions/repository/TransactionRepository.java
@@ -4,22 +4,19 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.ssor.boss.core.entity.Transaction;
 import org.ssor.boss.core.entity.TransactionType;
-
-import java.util.List;
 
 @Repository
 public interface TransactionRepository extends JpaRepository<Transaction, Integer>
 {
   @Query(value = "SELECT t FROM Transaction t WHERE t.accountId = :accountId AND t.id = :id")
-  Transaction findTransactionById(Integer id, Integer accountId);
+  Transaction findTransactionById(Integer id, Long accountId);
 
   @Query(value = "SELECT t FROM Transaction t " +
                  "WHERE t.accountId = :accountId " +
                  "AND lower(t.merchantName) LIKE lower(concat('%',:keyword,'%')) " +
                  "AND (:filter IS NULL OR :#{#filter.index()} = 0  OR :filter = t.type) ")
-  Page<Transaction> findTransactionsByAccountIdWithOptions(Integer accountId, String keyword, TransactionType filter, Pageable pageable);
+  Page<Transaction> findTransactionsByAccountIdWithOptions(Long accountId, String keyword, TransactionType filter, Pageable pageable);
 }
